+++ CHANGELOG.rst
@@ -6,6 +6,12 @@ dcicutils
 Change Log
 ----------
 
+8.5.0
+=====
+* Moved structured_data.py from smaht-portal to here; new portal_utils and data_readers modules.
+* Strip sheet name in data_readers.Excel; respecte (ignore) hidden sheets.
+
+
 8.4.0
 =====
 * More work related to SMaHT ingestion (bundle/sheet_utils, data_readers, etc).